[bc][log] Move log -> legacy/log log2 -> log
- existing codebase should stick with old version using commit
- log v2 is likely not useable, but it's better to just start building
new application/library using it
