+++ legacy/log/config_test.go
