+++ legacy/log/doc.go
