+++ legacy/log/entry.go
