+++ legacy/log/entry_generated.go
