+++ legacy/log/entry_test.go
