+++ legacy/log/filter.go
